Back/develop

See merge request s11-webmobile1-sub2/S11P12D103!159

+++ backend/src/main/java/com/mamdaero/domain/diary/service/DiaryService.java
@@ -3,17 +3,24 @@ package com.mamdaero.domain.diary.service;
 import com.mamdaero.domain.diary.dto.request.DiaryRequestDto;
 import com.mamdaero.domain.diary.dto.response.DiaryResponseDto;
 import com.mamdaero.domain.diary.entity.Diary;
+import com.mamdaero.domain.diary.exception.DiaryBadRequestException;
 import com.mamdaero.domain.diary.exception.DiaryNoContentException;
 import com.mamdaero.domain.diary.exception.DiaryNoDateException;
 import com.mamdaero.domain.diary.exception.DiaryNotFoundException;
 import com.mamdaero.domain.diary.repository.DiaryRepository;
 import com.mamdaero.domain.member.entity.Member;
+import com.mamdaero.domain.member.exception.MemberNotFoundException;
 import com.mamdaero.domain.member.repository.MemberRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageImpl;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 
 @Service
@@ -23,30 +30,52 @@ public class DiaryService {
     private final DiaryRepository diaryRepository;
     private final MemberRepository memberRepository;
 
-    public List<DiaryResponseDto> findAllByMember() {
+    public Page<DiaryResponseDto> findAllByMember(Long memberId, int page, int size) {
 
-        Member member = memberRepository.findById(1L).get();
+        Pageable pageable = PageRequest.of(page, size);
 
-        if (diaryRepository.findDiaryByMember(member).isEmpty()) {
-            throw new DiaryNotFoundException();
+        Optional<Member> optionalMember = memberRepository.findById(memberId);
+
+        if (optionalMember.isPresent()) {
+
+            Member member = optionalMember.get();
+
+            if (diaryRepository.findDiaryByMember(member, pageable).isEmpty()) {
+                throw new DiaryNotFoundException();
+            }
+
+            List<DiaryResponseDto> diaries = diaryRepository.findDiaryByMember(member, pageable).stream()
+                    .map(DiaryResponseDto::toDTO)
+                    .toList();
+
+            return new PageImpl<>(diaries);
         }
 
-        return diaryRepository.findDiaryByMember(member).stream()
-                .map(DiaryResponseDto::toDTO)
-                .toList();
+        throw new MemberNotFoundException();
     }
 
-    public List<DiaryResponseDto> findAllByMemberAndIsOpen(Boolean isOpen) {
+    public Page<DiaryResponseDto> findAllByMemberAndIsOpen(Long memberId, Boolean isOpen, int page, int size) {
 
-        Member member = memberRepository.findById(1L).get();
+        Pageable pageable = PageRequest.of(page, size);
 
-        if (diaryRepository.findAllByMemberAndIsOpen(member, isOpen).isEmpty()) {
-            throw new DiaryNotFoundException();
+        Optional<Member> optionalMember = memberRepository.findById(memberId);
+
+        if (optionalMember.isPresent()) {
+
+            Member member = optionalMember.get();
+
+            if (diaryRepository.findAllByMemberAndIsOpen(member, isOpen, pageable).isEmpty()) {
+                throw new DiaryNotFoundException();
+            }
+
+            List<DiaryResponseDto> diaryResponseDtoList = diaryRepository.findAllByMemberAndIsOpen(member, isOpen, pageable)
+                    .stream()
+                    .map(DiaryResponseDto::toDTO)
+                    .toList();
+
+            return new PageImpl<>(diaryResponseDtoList);
         }
-
-        return diaryRepository.findAllByMemberAndIsOpen(member, isOpen).stream()
-                .map(DiaryResponseDto::toDTO)
-                .toList();
+        throw new MemberNotFoundException();
     }
 
     public DiaryResponseDto findById(Long id) {
@@ -56,62 +85,73 @@ public class DiaryService {
         if (optionalDiary.isPresent()) {
             return DiaryResponseDto.toDTO(optionalDiary.get());
         }
-
         throw new DiaryNotFoundException();
     }
 
     @Transactional
-    public void create(DiaryRequestDto requestDto) {
+    public void create(Long memberId, DiaryRequestDto requestDto) {
 
-        Member member = memberRepository.findById(1L).get();
+        Optional<Member> optionalMember = memberRepository.findById(memberId);
 
-        if (requestDto.getContent().isEmpty()) {
-            throw new DiaryNoContentException();
+        if (optionalMember.isPresent()) {
+            Member member = optionalMember.get();
+
+            if (requestDto.getContent().isEmpty()) {
+                throw new DiaryNoContentException();
+            } else if (requestDto.getDate() == null) {
+                throw new DiaryNoDateException();
+            }
+
+            Diary diary = Diary.builder()
+                    .member(member)
+                    .content(requestDto.getContent())
+                    .date(requestDto.getDate())
+                    .isOpen(requestDto.getIsOpen())
+                    .build();
+
+            diaryRepository.save(diary);
+        } else {
+            throw new MemberNotFoundException();
         }
-        else if (requestDto.getDate() == null) {
-            throw new DiaryNoDateException();
-        }
-
-        Diary diary = Diary.builder()
-                .member(member)
-                .content(requestDto.getContent())
-                .date(requestDto.getDate())
-                .build();
-
-        diaryRepository.save(diary);
     }
 
     @Transactional
-    public void update(Long id, DiaryRequestDto requestDto) {
+    public void update(Long id, DiaryRequestDto requestDto, Long memberId) {
         Optional<Diary> optionalDiary = diaryRepository.findById(id);
 
         if (optionalDiary.isPresent()) {
             Diary diary = optionalDiary.get();
 
+            if (!Objects.equals(diary.getMember().getId(), memberId)) {
+
+                throw new DiaryBadRequestException();
+            }
+
             if (requestDto.getContent().isEmpty()) {
                 throw new DiaryNoContentException();
-            }
-            else if (requestDto.getDate() == null) {
+            } else if (requestDto.getDate() == null) {
                 throw new DiaryNoDateException();
             }
 
             diary.update(requestDto);
-        }
-        else {
+        } else {
             throw new DiaryNotFoundException();
         }
     }
 
     @Transactional
-    public void delete(Long id) {
+    public void delete(Long id, Long memberId) {
         Optional<Diary> optionalDiary = diaryRepository.findById(id);
 
         if (optionalDiary.isPresent()) {
             Diary diary = optionalDiary.get();
 
+            if (!Objects.equals(diary.getMember().getId(), memberId)) {
+                throw new DiaryBadRequestException();
+            }
+
             diaryRepository.delete(diary);
-        }
-        else {
+        } else {
             throw new DiaryNotFoundException();
         }
     }
